+++ src/main/java/com/easygame/sdk/repository/model/po/SampleGlaze.java
@@ -11,7 +11,7 @@ public class SampleGlaze {
     @NotBlank(message = "名称不能为空")
     private String name;
 
-    private Integer fineness;
+    private double fineness;
 
     private Date creation_date;
 
@@ -31,11 +31,11 @@ public class SampleGlaze {
         this.name = name;
     }
 
-    public Integer getFineness() {
+    public double getFineness() {
         return fineness;
     }
 
-    public void setFineness(Integer fineness) {
+    public void setFineness(double fineness) {
         this.fineness = fineness;
     }
